requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
prometheus-fastapi-instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
prometheus-fastapi-instrumentator
opentelemetry-instrumentation-sqlalchemy
pytest
